Ajustes Primer Version

+++ class/Producto.php
@@ -36,32 +36,31 @@ if(isset($_POST["action"])){
 class Producto{
     public $id=null;
     public $nombre='';
-    public $cantidad=0;
-    public $scancode='';
-    public $precio=0;
-    public $codigoRapido='';
-    public $idcategoria='';
-    public $fechaExpiracion=null;
+    public $nombreAbreviado='';
     public $descripcion='';
-    function __construct(){
-        require_once("Conexion.php");
-        //require_once("Log.php");
-        //require_once('Globals.php');
-        //
+    public $cantidad=0;    
+    public $precio=0;
+    public $scancode='';
+    public $codigoRapido='';
+    public $fechaExpiracion=null;
+    public $listacategoria= array();
+    
+    function __construct(){        
         // identificador único
         if(isset($_POST["id"])){
             $this->id= $_POST["id"];
         }
         if(isset($_POST["obj"])){
             $obj= json_decode($_POST["obj"],true);
-            $this->id= $obj["id"] ?? null;
+            require_once("UUID.php");
+            $this->id= $obj["id"] ?? UUID::v4();
             $this->nombre= $obj["nombre"] ?? '';
+            $this->nombreAbreviado= $obj["nombreAbreviado"] ?? '';
             $this->descripcion= $obj["descripcion"] ?? '';
-            $this->cantidad= $obj["cantidad"] ?? 0;            
-            $this->scancode= $obj["scancode"] ?? '';
+            $this->cantidad= $obj["cantidad"] ?? 0;                        
             $this->precio= $obj["precio"] ?? 0;
+            $this->scancode= $obj["scancode"] ?? '';
             $this->codigoRapido= $obj["codigoRapido"] ?? 0;
-            $this->idcategoria= $obj["idcategoria"] ?? null;
             $this->fechaExpiracion= $obj["fechaExpiracion"] ?? null;
             //Categorias del producto.
             if(isset($obj["listacategoria"] )){
@@ -78,7 +77,7 @@ class Producto{
     }
     function ReadAll(){
         try {
-            $sql='SELECT id, nombre, cantidad, scancode, cantidad, precio , codigoRapido
+            $sql='SELECT id, nombre, nombreAbreviado, scancode, cantidad, precio , codigoRapido
                 FROM     producto       
                 ORDER BY nombre asc';
             $data= DATA::Ejecutar($sql);
@@ -94,9 +93,10 @@ class Producto{
     }
     function Read(){
         try {
-            $sql='SELECT id, nombre, cantidad, scancode, precio , codigoRapido, idcategoria, fechaExpiracion, descripcion
-                FROM producto  
-                where id=:id';
+            $sql='SELECT p.id, p.nombre, p.nombreAbreviado, p.descripcion, cantidad, precio, scancode, codigoRapido, fechaExpiracion, c.id as idcategoria,c.nombre as nombrecategoria
+                FROM producto  p LEFT JOIN categoriasxproducto cp on cp.idproducto = p.id
+                    LEFT join categoria c on c.id = cp.idcategoria
+                where p.id=:id';
             $param= array(':id'=>$this->id);
             $data= DATA::Ejecutar($sql,$param);     
             foreach ($data as $key => $value){
@@ -155,11 +155,12 @@ class Producto{
     }
     function Create(){
         try {
-            $sql="INSERT INTO producto   (id,nombre, cantidad, scancode, precio ,codigoRapido, idcategoria, fechaExpiracion, descripcion)
-                VALUES (uuid(),:nombre, :cantidad, :scancode, :precio ,:codigoRapido, :idcategoria, :fechaExpiracion, :descripcion)";              
+            $sql="INSERT INTO producto   (id, nombre, nombreAbreviado, descripcion, cantidad, precio, scancode, codigoRapido, fechaExpiracion)
+                VALUES (:uuid, :nombre, :nombreAbreviado, :descripcion, :cantidad, :precio, :scancode, :codigoRapido, :fechaExpiracion)";
             //
-            $param= array(':nombre'=>$this->nombre,':cantidad'=>$this->cantidad,':scancode'=>$this->scancode, ':precio'=>$this->precio, ':codigoRapido'=>$this->codigoRapido, ':idcategoria'=>$this->idcategoria, ':fechaExpiracion'=>$this->fechaExpiracion, ':descripcion'=>$this->descripcion );
-            $data = DATA::Ejecutar($sql,$param,false);
+            $param= array(':uuid'=>$this->id, ':nombre'=>$this->nombre, ':nombreAbreviado'=>$this->nombreAbreviado, ':descripcion'=>$this->descripcion, ':cantidad'=>$this->cantidad, ':precio'=>$this->precio,
+                ':scancode'=>$this->scancode, ':codigoRapido'=>$this->codigoRapido, ':fechaExpiracion'=>$this->fechaExpiracion);
+            $data = DATA::Ejecutar($sql,$param, false);
             if($data)
             {
                 //save array obj
@@ -180,9 +181,10 @@ class Producto{
     function Update(){
         try {
             $sql="UPDATE producto 
-                SET nombre=:nombre, cantidad=:cantidad, scancode=:scancode, precio=:precio, codigoRapido=:codigoRapido, idcategoria=:idcategoria, fechaExpiracion=:fechaExpiracion, descripcion= :descripcion
+                SET nombre=:nombre, nombreAbreviado=:nombreAbreviado, descripcion= :descripcion, cantidad=:cantidad, precio=:precio, scancode=:scancode, codigoRapido=:codigoRapido, fechaExpiracion=:fechaExpiracion
                 WHERE id=:id";
-            $param= array(':id'=>$this->id, ':nombre'=>$this->nombre,':cantidad'=>$this->cantidad,':scancode'=>$this->scancode, ':precio'=>$this->precio , ':codigoRapido'=>$this->codigoRapido, ':idcategoria'=>$this->idcategoria, ':fechaExpiracion'=>$this->fechaExpiracion, ':descripcion'=>$this->descripcion );
+            $param= array(':id'=>$this->id, ':nombre'=>$this->nombre, ':nombreAbreviado'=>$this->nombreAbreviado, ':descripcion'=>$this->descripcion, ':cantidad'=>$this->cantidad, ':precio'=>$this->precio , 
+                ':scancode'=>$this->scancode, ':codigoRapido'=>$this->codigoRapido, ':fechaExpiracion'=>$this->fechaExpiracion);
             $data = DATA::Ejecutar($sql,$param,false);
             if($data){
                 //update array obj
@@ -206,8 +208,8 @@ class Producto{
                 'msg' => $e->getMessage()))
             );
         }
-    }   
-    function CheckRelatedItems(){
+    }  
+    private function CheckRelatedItems(){
         try{
             $sql="SELECT idproducto
                 FROM categoriasxproducto x
@@ -215,7 +217,7 @@ class Producto{
             $param= array(':id'=>$this->id);
             $data= DATA::Ejecutar($sql, $param);
             if(count($data))
-                return $data;
+                return true;
             else return false;
         }
         catch(Exception $e){
@@ -226,16 +228,17 @@ class Producto{
             );
         }
     }
+
     function Delete(){
         try {
-            // if($this->CheckRelatedItems()){
-            //     //$sessiondata array que devuelve si hay relaciones del objeto con otras tablas.
-            //     $sessiondata['status']=1; 
-            //     $sessiondata['msg']='Registro en uso'; 
-            //     return $sessiondata;           
-            // }                    
+            if($this->CheckRelatedItems()){
+                //$sessiondata array que devuelve si hay relaciones del objeto con otras tablas.
+                $sessiondata['status']=1; 
+                $sessiondata['msg']='Registro en uso'; 
+                return $sessiondata;           
+            }                    
             $sql='DELETE FROM producto  
-            WHERE id= :id';
+                WHERE id= :id';
             $param= array(':id'=>$this->id);
             $data= DATA::Ejecutar($sql, $param, false);
             if($data){
@@ -252,5 +255,6 @@ class Producto{
             );
         }
     }
+
 }
 ?>
